Saving work, added javadocs

+++ Inventory/src/main/java/wgu/softwareone/samircokic/inventory/controller/ModifyProductFormController.java
@@ -86,15 +86,18 @@ public class ModifyProductFormController implements Initializable {
 
     @FXML
     public void removeAssociatedPart(ActionEvent actionEvent) throws IndexOutOfBoundsException{
-        Alert alert = new Alert(Alert.AlertType.CONFIRMATION, "Do you want to delete this part?");
+        Alert alert = new Alert(Alert.AlertType.CONFIRMATION, "Do you want to remove this part?");
         Optional<ButtonType> answer = alert.showAndWait();
         try {
-            if (answer.isPresent() && answer.get() == ButtonType.OK) {
+            if (answer.isPresent() && answer.get() == ButtonType.OK&&partsInModifyProduct.size()>0) {
                 Part part = modifyProductAssociatedPartTable.getSelectionModel().getSelectedItem();
                 partsInModifyProduct.remove(part);
               Product product =  Inventory.lookupProduct(Integer.valueOf(modifyProductId.getText()));
                product.deleteAssociatedPart(part);
 
+            }else if( partsInModifyProduct.size() == 0){
+                Alert alertNotFound = new Alert(Alert.AlertType.ERROR, "Part not found");
+                alertNotFound.show();
             }
         } catch (IndexOutOfBoundsException e) {
             Alert alertNotFound = new Alert(Alert.AlertType.ERROR, "Part not found");
@@ -132,8 +135,11 @@ public class ModifyProductFormController implements Initializable {
                 scene = FXMLLoader.load(getClass().getResource("/wgu/softwareone/samircokic/inventory/MainMenu.fxml"));
                 stage.setScene(new Scene(scene));
                 stage.show();
-            }else {
-                Alert alert = new Alert(Alert.AlertType.ERROR, "Min should be less than Max; and Inv should be between those two values.");
+            }else if (min>max){
+                Alert alert = new Alert(Alert.AlertType.ERROR, "Min should be less than Max!");
+                alert.showAndWait();
+            }else if (inventory>max||inventory<min){
+                Alert alert = new Alert(Alert.AlertType.ERROR, "Inventory value should be between min and max values!");
                 alert.showAndWait();
             }
